pyproject.toml
[project]
name = "context-engine-demos"
version = "0.1.0"
description = "Context Engine Demos – multi-domain demo apps powered by Redis Context Surfaces."
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
  "context-surfaces>=0.0.1",
  "fastapi>=0.115.0",
  "httpx>=0.28.0",
  "langchain-openai>=1.1.12",
  "langgraph>=0.2.70",
  "langgraph-checkpoint-redis>=0.4.0",
  "openai>=1.68.0",
  "pydantic>=2.10.0",
  "pydantic-settings>=2.7.0",
  "python-dotenv>=1.0.1",
  "redis>=5.2.0",
  "redisvl>=0.16.0",
  "sentence-transformers>=3.0.0",
  "uvicorn[standard]>=0.34.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.25.0",
  "ruff>=0.8.0",
]

[tool.uv]
package = false

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
pythonpath = [".", "backend"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.8.0",
]
